Backfills basic 'getenvoy run' e2e test
